@@ -1,26 +1,23 @@
 source 'https://rubygems.org'
 
-gem 'rails', '~> 4.0'
+gem 'rails', '~> 5.0.0'
 gem 'rack'
 gem 'ransack'
 gem 'carrierwave'
-gem "pg", "~> 0.17.0"
+gem "pg"
 gem 'rmagick', '~> 2.15.4'
 gem 'simple_form'
 gem 'geocoder'
 gem 'cancan'
-# gem 'strong_parameters'
-# gem "taps"
-# gem 'validates_timeliness', '~> 3.0'
 
 #Handles authentication
-gem 'devise', '~> 3.0'
+gem 'devise', '>= 2.2.1'
 
 group :test, :development do
   gem 'thin'
-  gem "rspec-rails", '~> 3.4.0'
+  gem "rspec-rails"
   gem 'rspec-activemodel-mocks' # TODO: remove if we no longer need mock_model or stub_model
-  gem "factory_girl_rails", "~> 4.6"
+  gem "factory_girl_rails"
   gem "capybara"
   gem "guard-rspec", require: false
   gem 'guard-livereload'
@@ -31,7 +28,7 @@ group :test, :development do
   gem 'better_errors'
   gem 'binding_of_caller'
   gem 'meta_request'
-  gem 'database_cleaner', '1.0.1'
+  gem 'database_cleaner'
   gem 'test-unit'
   gem 'poltergeist'
   gem 'pry'
@@ -39,12 +36,18 @@ end
 
 group :test do
   gem 'shoulda-matchers'
+  # This is reportedly necessary to use the `assigns` and `assert_template` helpers
+  # in Rails 5.0.
+  gem 'rails-controller-testing'
 end
 
+group :assets do
+  # Gems used only for assets and not required
+  # in production environments by default.
   gem 'jquery-rails'
 
-  gem 'sass-rails',   '~> 4.0.0'
-  gem 'coffee-rails', '~> 4.0.1'
+  gem 'sass-rails'
+  gem 'coffee-rails'
   #The following provides the stylinmg for the datatables, among other things
   gem 'jquery-ui-rails'
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
@@ -52,17 +55,19 @@ end
 
   gem 'uglifier', '>= 1.0.3'
   gem 'jquery-datatables-rails', github: 'rweng/jquery-datatables-rails'
-  gem 'bootstrap-sass', '~> 2.0.4.1'
+  gem 'bootstrap-sass', '>= 2.0.4.1'
   gem 'chosen-rails'
+end
 
 # To use ActiveModel has_secure_password
-gem 'bcrypt-ruby', '~> 3.0.0'
+gem 'bcrypt-ruby', '>= 3.0.0'
 
 # To use Jbuilder templates for JSON
 # gem 'jbuilder'
-
-# Use unicorn as the app server
- gem 'unicorn'
+group :production do
+  # Use unicorn as the app server
+  gem 'unicorn'
+end
 
 # Deploy with Capistrano
 gem 'capistrano', '~> 2.15'
