@@ -1,5 +1,5 @@
 class SkillsController < ApplicationController
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
   load_and_authorize_resource
 
   # GET /skills
