@@ -1,5 +1,5 @@
 class TitlesController < ApplicationController
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
   load_and_authorize_resource
   # GET /titles
   # GET /titles.json
