@@ -1,4 +1,3 @@
-class Activity < ActiveRecord::Base
-  attr_accessible :author, :content, :loggable_id, :loggable_type
+class Activity < ApplicationRecord
   belongs_to :loggable, polymorphic: true
 end
