@@ -1,7 +1,5 @@
-class Event < ActiveRecord::Base
+class Event < ApplicationRecord
   before_save :calc_duration
-
-  attr_accessible :title, :description, :category, :course_id, :duration, :start_time, :end_time, :instructor, :location, :status, :timecard_ids, :person_ids, :comments
  
   def end_date_cannot_be_before_start
     if ((!end_time.blank?) and (!start_time.blank?)) and end_time < start_time
