@@ -1,3 +1,3 @@
-class Helpdoc < ActiveRecord::Base
-  attr_accessible :contents, :help_for_section, :help_for_view, :title
+class Helpdoc < ApplicationRecord
+  # attr_accessible :contents, :help_for_section, :help_for_view, :title
 end
