@@ -1,4 +1,4 @@
-class Inspection < ActiveRecord::Base
+class Inspection < ApplicationRecord
 
   belongs_to :item
   has_many :inspection_questions
