@@ -1,4 +1,4 @@
-class InspectionQuestion < ActiveRecord::Base
+class InspectionQuestion < ApplicationRecord
   belongs_to :inspection
   belongs_to :question
 end
