@@ -1,5 +1,5 @@
-class ItemType < ActiveRecord::Base
-  attr_accessible :is_a_group, :is_groupable, :name, :parent_id, :status
+class ItemType < ApplicationRecord
+  # attr_accessible :is_a_group, :is_groupable, :name, :parent_id, :status
   has_many :items
   def to_s
     name
