@@ -1,6 +1,6 @@
-class Location < ActiveRecord::Base
-  attr_accessible :category, :comments, :description, :department_id,
-                  :floor, :container, :lat, :lon, :name, :status
+class Location < ApplicationRecord
+  # attr_accessible :category, :comments, :description, :department_id,
+  #                 :floor, :container, :lat, :lon, :name, :status
   has_many :moves
   belongs_to :department
   def to_s
