@@ -1,3 +1,3 @@
-class Move < ActiveRecord::Base
-  attr_accessible :comments, :item_id, :locatable_id, :locatable_type, :reason
+class Move < ApplicationRecord
+  # attr_accessible :comments, :item_id, :locatable_id, :locatable_type, :reason
 end
