@@ -1,5 +1,5 @@
-class Role < ActiveRecord::Base
-  attr_accessible :name
+class Role < ApplicationRecord
+  # attr_accessible :name
   has_and_belongs_to_many :users
 
   def to_s
