@@ -1,5 +1,5 @@
-class Skill < ActiveRecord::Base
-  attr_accessible :name, :status, :course_ids, :title_ids
+class Skill < ApplicationRecord
+  # attr_accessible :name, :status, :course_ids, :title_ids
   has_and_belongs_to_many :courses
   has_and_belongs_to_many :titles
   validates_presence_of :name, :status
