@@ -1,5 +1,5 @@
-class Title < ActiveRecord::Base
-  attr_accessible :comments, :description, :status, :name, :skill_ids
+class Title < ApplicationRecord
+  # attr_accessible :comments, :description, :status, :name, :skill_ids
     has_and_belongs_to_many :people
     has_and_belongs_to_many :skills
 
