@@ -1,6 +1,6 @@
-class UniqueId < ActiveRecord::Base
+class UniqueId < ApplicationRecord
   belongs_to :item
-  attr_accessible :category, :status, :value, :item
+  # attr_accessible :category, :status, :value, :item
 
   CATEGORY_CHOICES = ['VIN', 'Sub-component', 'Grant', 'Slug', 'Registration',
                       'Department', 'Other']
