@@ -37,7 +37,7 @@ describe ActivitiesController do
   describe "GET index" do
     it "assigns all activities as @activities" do
       activity = Activity.create! valid_attributes
-      get :index, {}, valid_session
+      process :index, method :get, {}, valid_session
       expect(assigns(:activities)).to eq([activity])
     end
   end
@@ -45,14 +45,14 @@ describe ActivitiesController do
   describe "GET show" do
     it "assigns the requested activity as @activity" do
       activity = Activity.create! valid_attributes
-      get :show, {:id => activity.to_param}, valid_session
+      process :show, method :get, {:id => activity.to_param}, valid_session
       expect(assigns(:activity)).to eq(activity)
     end
   end
 
   describe "GET new" do
     it "assigns a new activity as @activity" do
-      get :new, {}, valid_session
+      process :new, method :get, {}, valid_session
       expect(assigns(:activity)).to be_a_new(Activity)
     end
   end
@@ -60,7 +60,7 @@ describe ActivitiesController do
   describe "GET edit" do
     it "assigns the requested activity as @activity" do
       activity = Activity.create! valid_attributes
-      get :edit, {:id => activity.to_param}, valid_session
+      process :edit, method :get, {:id => activity.to_param}, valid_session
       expect(assigns(:activity)).to eq(activity)
     end
   end
@@ -69,18 +69,18 @@ describe ActivitiesController do
     describe "with valid params" do
       it "creates a new Activity" do
         expect {
-          post :create, {:activity => valid_attributes}, valid_session
+          process :create, method :post, {:activity => valid_attributes}, valid_session
         }.to change(Activity, :count).by(1)
       end
 
       it "assigns a newly created activity as @activity" do
-        post :create, {:activity => valid_attributes}, valid_session
+        process :create, method :post, {:activity => valid_attributes}, valid_session
         expect(assigns(:activity)).to be_a(Activity)
         expect(assigns(:activity)).to be_persisted
       end
 
       it "redirects to the created activity" do
-        post :create, {:activity => valid_attributes}, valid_session
+        process :create, method :post, {:activity => valid_attributes}, valid_session
         expect(response).to redirect_to(Activity.last)
       end
     end
@@ -89,14 +89,14 @@ describe ActivitiesController do
       it "assigns a newly created but unsaved activity as @activity" do
         # Trigger the behavior that occurs when invalid params are submitted
         allow_any_instance_of(Activity).to receive(:save).and_return(false)
-        post :create, {:activity => { "content" => "invalid value" }}, valid_session
+        process :create, method :post, {:activity => { "content" => "invalid value" }}, valid_session
         expect(assigns(:activity)).to be_a_new(Activity)
       end
 
       it "re-renders the 'new' template" do
         # Trigger the behavior that occurs when invalid params are submitted
         allow_any_instance_of(Activity).to receive(:save).and_return(false)
-        post :create, {:activity => { "content" => "invalid value" }}, valid_session
+        process :create, method :post, {:activity => { "content" => "invalid value" }}, valid_session
         expect(response).to render_template("new")
       end
     end
